Rename command and folder
